Multiple event builders per device
The thing is that sometimes we have fragments
from different events mixed. If we have only
one event builder per device then when we get
fragment from next event we have to close
current event even if we haven't got all the fragments.

And when we close the event we can not then restore
it even if we eventually get necessary fragment.